Avoid a documentation 401
The docs currently contain links to the 'master-SNAPSHOT' tree, which appears to require
authorization. The 'master' tree does not. This change naively replaces all 'master-SNAPSHOT'
links with the corresponding links in the 'master' tree.

I have not checked all the links for validity, and I don't understand the build enough to
know if this is the right place to point to. I would imagine that probably the docs should
point to the latest-stable release by default, such as the '6.11' tree.